Fix clang-tidy errors
Add INDEX_BITS

+++ src/backends/base/mfe/mfe_debug.cpp
@@ -48,7 +48,7 @@ void MfeDebug::Run(const Primary& r, const Model::Ptr& m, DpState& state, erg::E
 
   logdebug("base {} with {}, {}", funcname(), cfg, pf);
 
-  const int N = static_cast<int>(r.size());
+  const int N = r.size();
   state.dp = DpArray(r.size() + 1, MAX_E);
   auto& dp = state.dp;
