+++ src/backends/base/pfn/pfn_opt.cpp
@@ -45,7 +45,7 @@ PfnTables PfnOpt::Run(const Primary& r, const BoltzModel::Ptr& bm, erg::EnergyCf
 
   logdebug("base {} with {}, {}", funcname(), cfg, pf);
 
-  const int N = static_cast<int>(r.size());
+  const int N = r.size();
   const BoltzPrecomp bpc(Primary(r), bm, cfg, pf);
   state.dp = BoltzDpArray(r.size() + 1, 0);
   auto& dp = state.dp;
